requirements.txt
streamlit==1.45.0
google-auth-oauthlib==1.0.0
google-auth==2.25.0
requests==2.31.0
plotly==5.18.0
urllib3==2.1.0